// src/app/interfaces/club.ts
export interface Club {
  id?: string;
  altNameEn: string;    // "Red Devils"
  altNameRu: string;    // "Красные Дьяволы"
  budget: number;       // 100
  league: string;       // "/leagues/NIDSxlgrIPnZUoKqiw10"
  leagueNameEn: string; // "Premier League"
  leagueNameRu: string; // "Премьер Лига"
  nameEn: string;       // "Manchester United"
  nameRu: string;       // "Манчестер Юнайтед"
  stadium: number;      // 80000
}